+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Added: Add attribute `role` to the flyout block element to `dialog` or `alertdialog` with support to set the value via `data-flyout-role`.
 - Added: Add support to the attribute `hidden` on flyout block elements to hide content for when JavaScript and CSS is unavailable.
 - Added: Support to target flyout blocks using the `href` of a link attribute.
 - Added: Set trigger elements `role` to `button`.